Allow adding sub-headings under Proposed change header
This changes the titles test under the "Proposed change" heading so
that it checks that *at least* the required 9 headings exist instead
of checking that *exactly* the required 9 headings exist.

This will allow writers like me to add some structure to the first
section under this heading before getting to the required sections.

Change-Id: I507d32cb2568d9c65b7a460196fdd873c4c5c445