Extension for Network Security Group
Neutron offers virtual network management functionality.
Since we can define any number of networks/subnets in a flexible way,
so virtual network topologies
tend to map security policy. On the other hand, we can specify security group
on a per port basis.
In this blueprint, we are proposing the addition of security_groups
as an extension attribute for the Network resource, for better API usability.
As an example, with this extension a tenant can define network-wide default security group(s),
so any vms that come up on any subnet in that network will be subject to the corresponding group rules.

Change-Id: I760f5d99d17edbdc4ebf9752869f9e1f3a8d3a21