Service group and service object for FWaaS blueprint fwaas-customized-service
In the traditional firewall design a service is used to define type of traffic
in firewall. This blueprint creates an extension that allows the firewall
administrators to create customized service objects. The customized service
objects can be grouped together to form a service group object. The policy
rule of FWaaS can reference more than one type of traffic by referencing to
service group objects.

Change-Id: I7a05799a2f8e51438852267cd62d1030ae74ceab