Neutron OVS agent - Dont Fragment flag
This blueprint spec provides information
about a new option for the OVS agent.

It introduces a dont_fragment option that
allow OVS tunnel -GRE and VXLAN- to be
fragmented at the IP layer.

It will permit to carry virtual network
with a bigger (or equal) MTU than the
underlying physical network.

Change-Id: I5a490469f270bbcd014f3f8fb55cf024ff7f3e00