Add spec for retargetable functional testing
This spec proposes the addition of support for retargetable
functional testing.  Retargetable tests are intended to be written
to an abstract client interface that can have multiple backends,
allowing an API test to be written that can be run against the
plugin API and a Neutron service with minimal extra effort.

bp retargetable-functional-testing

Change-Id: I1aa1e147012f853d3c192de80505e6ce0d50b5c4