Service group and Service Object for firewall as a service
In the traditional firewall design a service is used to define type of traffic
in firewall. This blueprint creates an extension that allows the firewall
administrators to create customized service objects. The customized service
objects can be grouped together to form a service group object. The policy
rule of FWaaS can reference more than one type of traffic by referencing to
service group objects.
Implements: blueprint fwaas-customized-service for customized service
Change-Id: I4d126a313017ef01308d565b6792e104a5d1a935
Authored-by: badveli vishnuvardhan <badveli_vishnuus@yahoo.com>