Proposal for Neutron core and vendor code decomposition
This spec outlines the plan for addressing the pain points
that every Neutron contributor feels on a daily basis when
working with the project.

The proposed change should be implemented collaboratively,
in the same spirit of the OpenStack community.

This is initial attempt at making this project a thriving
one. We should all recognize that if nothing gets done,
Neutron will become a desolated land where just tumbleweeds
will abound.

Blueprint: core-vendor-decomposition

Change-Id: I7156edbeeed8d5e11415d3feb5b340bbe74e327f