Filtering routers by network-id
There is currently no way to list all the routers who are connected to
a single network. This patch adds an optional 'network_id' filter to
the router-list REST API, which (when used) will allow one to get a
list of all the routers which belong to a specific network.

bp filter-routers-by-network-id

Change-Id: I8d0a02a6f214df9162760a5207a239703033d82c