MTU selection spec update bridge_mappings new format
Update the spec with the revision of the agreed
per bridge_mappings/physnet MTU setting format.  This
removes reference to the bridge_mappings config as it
is per-instance of compute-node agent, thus, a scale
and usability problem to use that as a vehicle for
providing this info.  The implementation in review
 https://review.openstack.org/153733/
will need to be revised to match the new proposal.

blueprint mtu-selection-and-advertisement

Change-Id: I7bfe4fc5c58a3894785835f028b3fc7e8dcb2f9c