Shorten the spec template greatly.
Since neutron is moving to using RFE's for features, only the larger
or contentious features will require specs. This change removes many
of the spec sections that will become redundant with the information
included in the RFE itself and the devref developer docs that are now
expected with the code.

Also, tweak the py27 test to accept both old and new spec formats

Change-Id: I5aeb2cb39740ac85e0485f4e26d6fb380bd4d271