"Get me a network" spec
Neutron has a lot of concepts that some users don't want or need to
understand, but must be configured if 'nova boot' is to succeed in
a Neutron-based deployment. For this reason, the bar has been set
a little too high for those users migrating from a Nova-network based
deployment (where network are pre-defined by the admin), to a Neutron-
based one, where users get to define their own networks.

Documenting how administrators can ease this transition is required,
this spec is a step at addressing this, and outlines the options and
some possible future enhancements for Neutron.

Change-Id: Id743b8989cab3fb13cf6f079acc4dedf80039e19