Specs cleanup for the Liberty release
This moves most of the Liberty specs which were not implemented into the
liberty-backlog directory. However, two specs were moved directly to
Mitaka:

external-dns-resolution.rst
restructure-l2-agent.rst

Both of those are pre-approved for Mitaka based on the work that has been
going on in Liberty (and Kilo in the case of restructure-l2-agent.rst).

Change-Id: I4d7d48ad96d01e9013fc01ea6573208b63f16521
Signed-off-by: Kyle Mestery <mestery@mestery.com>