Nuke project priorities
It is too difficult and time consuming to track this by hand and
the section essentially duplicates information that's avaialable
in Launchpad.

The idea of highlighting who does what in a cycle is nice, but
this needs some more thinking.

For now, kill the section, which strikes as odd amongst the
other content of the landing page.

Change-Id: I78f18ed8418f01599fd6acb58317c70bb7f633e1