Neutron Common Classification Framework
This patch documents the Common Classification Framework.

Neutron features/projects that may need classification are:
- Security group rules
- openstack/neutron-fwaas
- openstack/networking-sfc
- openstack/networking-bgpvpn
- openstack/tap-as-a-service
- Neutron QoS

Furthermore, there are other projects with classication APIs, such as
openstack/group-based-policy and it's possible that others will want
to support classifications in their own APIs, further reinventing the wheel
and fragmenting the language used in the OpenStack ecosystem when it comes
to defining traffic classifications.

Change-Id: I6df22d9e766d8aeabc2f99262cadeec332d809b0
Related-Bug: #1476527