Layer 3 IPs rate limit
The patch proposes a general solution for the L3 IP QoS,which adds
`qos_policy_id` to the L3 IPs. Then the QoS rules will be used to
set the linux tc rules in router namespace efficiently. For the L3
users, the new attribute would indicate the bandwidth they can use.

Change-Id: If7dba69857a6758ef278905cd10434b1de4447b5
Related-Bug: #1596611