Repropose: Spec for providing Nova portbinding information for live migration
This patch is reproposal of change proposed by Andreas at
I12d1310b8df30f5fbae4cb9c8e804d34c20ee7c9 planned for Ocata.

The goal is to provide nova the required portbinding information
for live migration. This spec proposes to allow a port to have
a second inactive binding. During migration, the host that
runs the instance has the active binding, while the binding
belonging to the other host is inactive.

A new API is required to externalize this new inactive binding.

Partial-Bug: #1580880
Co-authored-by: Brian Stajkowski <brian.stajkowski@rackspace.com>

Change-Id: I707e48f6c13f80f2696ac877c9e9948eb453e6e2