decouple neutron db models and ovo for neutron-lib
This spec provides details on how we can break consumer dependencies
on neutron's database model as well as object imports by
exposing neutron objects as entry points that are discoverable and
loadable from neutron-lib. This spec is part of a set of specs related
to decoupling the db for neutron-lib (see patch topic).

Change-Id: I079d06502e6e7b1e20aea882979b0ecd9106eaeb