Fix py27 job in specs repo
In py27 job index.rst file was checked as any other file if it
contains required headers and sections.
But index.rst file has different format so this file should be
skipped in this test.

After that there was also issue with unexpected "Footnotes"
section in neutron-inter.rst specs so this was also fixed.

Change-Id: If2baf887c49025a5481e97b2fb3500736b8668b6