Add list of all working DSCP marks
The QoS DSCP spec is not terribly specific about what DSCP marks are
valid.  This change makes it explicit.

This spec is quite old.  But since there has been no other place in the
documentation that explicitly lists what the valid DSCP marks are,
updating this in case people use it as documentation.

Change-Id: Ieec3e238b8d404e46c49b5a1a0fdaf695fff4252
Related-Bug: #1781915