Network Segment Range Management
Currently, network segment ranges are configured as an entry in ML2
config file [1]_ that is statically defined for tenant network
allocation and therefore must be managed as part of the host deployment
and management. When a normal tenant user creates a network, Neutron
assigns the next free segmentation ID (VLAN ID, VNI etc.) from the
configured segment ranges; only an administrator can assign a specific
segment ID via the provider extension.

This spec introduces an extension which exposes the segment range
management to be administered via the Neutron API. In addition, it
introduces the ability for the administrator to control the segment
ranges globally or on a per-tenant basis.

[1] /etc/neutron/plugins/ml2/ml2_conf.ini

Co-Authored-By: Matt Peters <matt.peters@windriver.com>

Implements: blueprint network-segment-range-management
Change-Id: I2e586bb6d51d32c0ac2a28b2429512cea036f363