Changing segmentation ID of existing network should be allowed
In the context of a live datacenter, sometimes the network teams re-organize
their VLANs by changing VLAN IDs. This is pretty straightforward in the
context of network configuration and physical servers, but when using ML2
VLAN provider networks, those IDs cannot be updated. The only possible
solution here would be to delete those networks and recreate them with a
different segmentation ID which might be a problem when you have 1000 VMs
and 2000 ports.

For this reason it would be good if we could change the segmentation ID of
an existing network.

Change-Id: I71b5b4ff2ba7ace3a8523de2ae6e4dc6a1c110c2
Related-Bug: #1806052