Port binding event extended information for Nova
There are several cases in the Nova/Neutron/os-vif interaction
where the knowledge of the neutron core plugin or ML2 driver
would be useful to facilitate a more robust handling of guest
networking.

For example, in Icehouse, a external event mechanism was
introduced to allow neutron to notify Nova of network changes
and VIF plugging. As the occasion on which these events are
emitted is dependent on the driver that bound the port and Nova
does not have knowledge of that. Therefore, Nova cannot rely on
the event to synchronize the operation between Nova and neutron,
which has lead to a race condition in the past.

Change-Id: I57f4df31bee83b818f2c0123577d11a8e07c0191
Related-Bug: #1821058