Port NUMA affinity policy, revisit
Since the conversation reported in [1], the information to be reported
to Nova will not be stored in the port "resource_request" and the
NUMA affinity policies will not be considered as resource provider
traits.

NUMA is going to be modeled in Placement soon, but until this feature
does not land in a stable version, Neutron will only provide this
information via an extension.

Related-Bug: #1886798

[1]https://review.opendev.org/#/c/739961/

Change-Id: I2a96e5ba394f4476121cbc916634e1c4d4b4212f