QoS minimum guaranteed packet rate
Similarly to how bandwidth can be a limiting factor of a network
interface, packet processing capacity tend to be a limiting factor
of the soft switching solutions like OVS. In the same time certain
applications are dependent on not just guaranteed bandwidth but also
on guaranteed packet rate to function properly. OpenStack already
supports bandwidth guarantees via the minimum bandwidth QoS policy
rules. This specification is aiming for adding support for a similar
minimum packet rate QoS policy rule.

Related-Bug: 1922237
Change-Id: Id17ee01dc288517a05f746a479500a6218ad55f4