Create intermediate OVS bridge to improve live-migration in OVN
This spec proposes to add an intermediate bridge between the VM patch
port and the integration bridge. That will allow the backend (OVN)
to properly configure the needed OpenFlow rules before the VM
is unpaused in the destination host. That will reduce the
networking disruption during the live migration process.

Change-Id: I558523a8922567efb0739173c7c2fda72504a8fe
Related-Bug: #1933517