Use VNIC_REMOTE_MANAGED instead of VNIC_SMARTNIC
After a round of reviews of Nova patches that utilized VNIC_SMARTNIC for
the off-path backend spec

https://review.opendev.org/q/topic:2021-09-10-off-path-net-backends-dep

it was determined that the Ironic's usage of VNIC_SMARTNIC would be
affected by how the decision is made whether a port is remote-managed or
not during the resource request creation.

https://review.opendev.org/c/openstack/nova/+/824835/13/nova/network/neutron.py#2325

See the following log for the relevant discussion:

https://meetings.opendev.org/irclogs/%23openstack-nova/%23openstack-nova.2022-02-07.log.html#t2022-02-07T15:49:56

Therefore, this change introduces a spec update to use a new VNIC
type.

Nova spec: https://review.opendev.org/c/openstack/nova-specs/+/787458
Nova BP: https://blueprints.launchpad.net/nova/+spec/integration-with-off-path-network-backends

Partial-Bug: #1932154
Change-Id: I63f156c5bfb5a41e5ebf94dc8f069828569e270a