Update envlist for tox4
tox4 even when running specific env with -e option tries
to find python version for all envs defined in envlist[1]
when running with --skip-missing-interpreters=false and
fails if interpreter missing for any of the env.

With py38 in envlist it tries to find python3.8
which is not available in ubuntu-jammy by default
and fails, hence use generic version instead i.e py3
to handle it.

Even though [1] is fixed we are still seeing this failure.

[1] https://github.com/tox-dev/tox/issues/2811

Also had to add "ignore_basepython_conflict = True".

Change-Id: I351870695b1c6706888d8a51fa33ea0fcbc63cd7