Update to use openstack-python3-jobs template
Using openstack-python3-jobs will allow this to
seamlessly move forward with the python version,
which currently should be py39 and py312.

Removed the py38 override as it will not be used
any more, and for reference, the template does
contain a similar irrelevant-files line so that
is not lost.

TrivialFix

Change-Id: Ieffe3581dfa444ae5d014d4b499c3ee7694e7e57